chore: renaming audio cnn model
