+++ src/core/operators/audio_cnn_model/config.py
