+++ src/core/operators/audio_cnn_model/inference.py
