+++ src/core/operators/audio_cnn_model/models.py
